requirements.txt
fastapi==0.115.2
uvicorn==0.32.1
requests==2.32.3
python-dotenv==1.0.1
supabase==2.5.1
python-telegram-bot==20.8
google-api-python-client==2.186.0
google-auth==2.41.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
